// src/renderer/src/store/model/type.d.ts
export type ModelConfig = Pick<IClientConfig, 'apiKey' | 'baseUrl' | 'mode' | 'id' | 'options'> & { model: string }

type BaseCompletionOptions = {
  signal?: AbortSignal
  enable_search?: boolean
  timeout?: number
  max_tokens?: number
}
export type StreamOptions = BaseCompletionOptions & {
  onError?: (code: string, message: string, e?: Error) => void
  onFinish?: (content: string) => void
  onChunk?: (fullText: string, text: string) => void
  onReasoning?: (reasoning: string) => void
}

export type CompletionOptions = BaseCompletionOptions & {}